Initial NavElements implementation
